# Conflicts:
#	imgui.ini

+++ Source/Game/Piece.cpp
@@ -20,9 +20,9 @@ Piece::~Piece()
 void Piece::Initialize(int index)
 {
 	//	駒の座標 = (将棋盤(9x9マス)上での座標 + pieceOffset) * range_
-	GetTransform()->SetPositionX((pieceInfo_[index].posX_ + pieceOffset_.x) * range_);
+	GetTransform()->SetPositionX((static_cast<float>(pieceInfo_[index].posX_) + pieceOffset_.x) * range_);
 	GetTransform()->SetPositionY(Stage::Instance().GetTransform()->GetPosition().y);	//将棋盤と高さ合わせるため補正とかしない
-	GetTransform()->SetPositionZ((pieceInfo_[index].posY_ + pieceOffset_.z) * range_);
+	GetTransform()->SetPositionZ((static_cast<float>(pieceInfo_[index].posY_) + pieceOffset_.z) * range_);
 
 	//	敵のときだけモデルの向きを反転させる
 	if (pieceInfo_[index].isEnemy_)GetTransform()->SetRotationY(DirectX::XMConvertToRadians(180));
